refactor: rename Template's .retrieve() to .path()
This member returns the path to the template file so it makes more sense
to call it just that. Additionally, this new name hides the fact that
the call might need to retrieve a local copy of the template file, which
is not always the case and really depends on internals of the fsource,
that should not be exposed.